refactor: reorganize files, improve typing, and update utilities
- Add type annotations to train_batch_end callback method
- Switch from cKDTree to KDTree in utils_grid
- Add short descriptive text to visualization(plot_loss=True)
- Move callbacks.py to callbacks/ folder and loss_function.py to metrics/ folder
